engine + point-of-entry: CI is made to tell the truth, and three of the four failures were never this branch's -- MASTER ALREADY FAILED clang-tidy, lizard and cppcheck before any of this existed, which is the first thing I should have checked and the reason a red pipeline on a new branch is not evidence about the new branch; only Format was mine, from includes inserted at the top of files where clang-format wanted them sorted; THE TWO ANALYSIS JOBS COULD NOT HAVE PASSED WHATEVER THE CODE SAID: both read compile_commands.json and the job comment asserts that configure alone is enough because the dependencies are FETCHED there -- but glad GENERATES glad.h and glad.c at BUILD time, so clang-tidy met twelve file-not-found errors before a single check ran and cppcheck could not load the project at all; one [0/2] Re-checking globbed directories... ninja: no work to do. in each is the difference between a job reading the code and a job failing on its own setup, and the false claim in the comment is corrected rather than left to mislead the next person; LIZARD WAS UNPINNED, so CI installed 1.24.0 while local had 1.21.2 and the two disagreed about whether the tree was clean -- pinned to what CI actually runs rather than to what happened to be installed here, since the newer one is the stricter and it was RIGHT: emitTileQuad has nine parameters against a documented limit of eight and the old version simply did not say so, which also means my earlier 'point-of-entry is at zero warnings' was an artefact of the tool rather than a fact about the code; the atlas facts are one struct now and RenderSystem::render gives up its rim-outline pass, a whole second draw with its own uniforms and UV arithmetic that the entity loop had to be read through; and the nine REAL clang-tidy findings in point-of-entry are gone -- two float loop counters where the accumulated error decided how many steps you got, a rounding cast that is lround now, a narrowing int, a const, an unused parameter, and three functions over the complexity threshold, each split along a seam it already had: what is still SURFACING out of chase, the live SHEET out of the debug panel, and the brew and points PAGES out of the staging screen behind a layout struct, so a row is hit-tested against exactly the rect it was drawn into

+++ games/point-of-entry/src/ecs/BalanceConfig.cpp
@@ -139,7 +139,8 @@ void applyDerivations(EntityManager& em, entt::entity entity)
     const float hpFrac =
         hp.max > 0 ? static_cast<float>(hp.current) / static_cast<float>(hp.max) : 1.0f;
     hp.max = newMaxHp;
-    hp.current = std::min(newMaxHp, static_cast<int>(std::lround(hpFrac * newMaxHp)));
+    hp.current =
+        std::min(newMaxHp, static_cast<int>(std::lround(hpFrac * static_cast<float>(newMaxHp))));
 
     const float newMaxSta = maxStamina(s);
     auto& sta = reg.get_or_emplace<Stamina>(entity, Stamina{newMaxSta, newMaxSta});
